Decoding multiple instructions and suffixes
This is not the most beautiful code.
I'm still clumsy with Zig.
The code is full of if-else statements. Sometimes nested.
I've tried to experiment with different approaches.
I think like more to create a small function for each case instead of
having one big function with a bunch of conditions in it. Even though
the small functions are duplicated somewhat.

Tested with nasm + cmp. No differences with the original listings.
Did both listing_39 and listing_40.