Add extractBits function
We have the sizes of bits, we have the sequence of sizes of bits and
we have the bytes buffer. Now we need to extract those bits from the
correct place in the buffer.

I'm not super happy with the funciton's APi. Maybe I will think of
something better when I try to actually use it. At least I have tests
for it.