Decode many instructions in a row
Now we have not just 2 bytes but a whole buffer of bytes (fine, it's
only six). So we need to make sure we extract all the instructions
from them bytes.

Why do I use "we"? Is it my computer and I? It's definitely not a
royal we and definitely not several personalities in one brain we. So
I guess it's the computer thing. Or I'm just being used to this kind
of wording from all that teamwork.