Trying specialized decoder functions
The `decodeOperand` function is getting more and more complex. I was
trying to keep everything in one function to minimize the code
duplication but it's getting harder and harder to keep track of which
condition is trying to cover which opcode.

With this commit I'm trying to move one opcode decoding out into a
separate function. Seems to work. I'll try a couple more to see how it
goes.

TLDR; It's ok to have duplicated code if it's easier to maintain/read.