Fix mov to memory
All that refactoring on this beautiful rainy day so that I could
reason about mov to memory instruction.
After the refactoring no new code was needed, just the correctly
executed funciton that I've had already.
Interestingly enough it doesn't matter where I pu the size specifier.
At least for NASM in the context of mov immediate to memory. So I
won't bother for longer to make sure I render the dissassembly exactly
as input, since the output binary is the same.
Now I can work on simulating the memory!