Big refactoring and some clocks
The refactoring is to simplify the different mem calculation structs.
I've moved from trying to share the data structures to have three
specific ones: MemCalcWithDisp, MemCalcNoDisp and DirectAddress.

Had to update a lot of parts of the code because of this but it
removed the duplicated DirectAddress and made things a bit more
clearer.

Also started to add Clocks to the instructions.